Multi-Container-Docker/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
jaraco.context>=6.1.0
prometheus-client